+++ test/test_util.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 from collections.abc import Iterator
 import json
-from pathlib import Path
-import shutil
 import time
 from typing import Optional
 from packaging.specifiers import SpecifierSet
 import pytest
 from pytest_mock import MockerFixture
-from pyrepo.commands.release import advance_devstatus, get_mime_type
+from pyrepo.commands.release import get_mime_type
 from pyrepo.util import (
     Bump,
     PyVersion,
@@ -18,7 +16,6 @@ from pyrepo.util import (
     sort_specifier,
     update_years2str,
 )
-from test_helpers import case_dirs
 
 
 @pytest.mark.parametrize(
@@ -174,15 +171,6 @@ def test_bump_version_prerelease(v: str, bump: Bump) -> None:
     with pytest.raises(ValueError) as excinfo:
         bump_version(v, bump)
     assert str(excinfo.value) == f"Cannot bump pre-release versions: {v!r}"
-
-
-@case_dirs("advance_devstatus")
-def test_advance_devstatus(dirpath: Path, tmp_path: Path) -> None:
-    shutil.copyfile(dirpath / "before.cfg", tmp_path / "setup.cfg")
-    advance_devstatus(tmp_path / "setup.cfg")
-    assert (tmp_path / "setup.cfg").read_text(encoding="utf-8") == (
-        dirpath / "after.cfg"
-    ).read_text(encoding="utf-8")
 
 
 @pytest.mark.parametrize(